+++ src/main/java/com/example/fastcampusmysql/domain/post/entity/Post.java
@@ -2,19 +2,25 @@ package com.example.fastcampusmysql.domain.post.entity;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+import javax.persistence.*;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Getter
 @Builder
+@Entity
+@NoArgsConstructor
+@Table(name = "POST")
 public class Post {
-    private final Long id;
-    private final Long memberId;
-    private final String contents;
-    private final LocalDate createdDate;
-    private final LocalDateTime createdAt;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private Long memberId;
+    private String contents;
+    private LocalDate createdDate;
+    private LocalDateTime createdAt;
 
     public Post(Long id, Long memberId, String contents, LocalDate createdDate, LocalDateTime createdAt) {
         this.id = id;
