+++ src/main/java/com/example/fastcampusmysql/domain/post/entity/Timeline.java
@@ -2,16 +2,24 @@ package com.example.fastcampusmysql.domain.post.entity;
 
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
 import java.time.LocalDateTime;
 import java.util.Objects;
 
 @Getter
+@Entity
+@NoArgsConstructor
 public class Timeline {
-    private final Long id;
-    private final Long memberId;
-    private final Long postId;
-    private final LocalDateTime createdAt;
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private Long memberId;
+    private Long postId;
+    private LocalDateTime createdAt;
 
     @Builder
     public Timeline(Long id, Long memberId, Long postId, LocalDateTime createdAt) {
