+++ src/main/java/com/example/fastcampusmysql/domain/post/repository/PostRepository.java
@@ -1,186 +1,23 @@
 package com.example.fastcampusmysql.domain.post.repository;
 
-import com.example.fastcampusmysql.util.PageHelper;
 import com.example.fastcampusmysql.domain.post.dto.DailyPostCount;
 import com.example.fastcampusmysql.domain.post.dto.DailyPostCountRequest;
 import com.example.fastcampusmysql.domain.post.entity.Post;
-import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 
-import java.sql.ResultSet;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 
-@Repository
-@RequiredArgsConstructor
-public class PostRepository {
-    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
-    private static final String TABLE = "POST";
-    private static final RowMapper<Post> ROW_MAPPER = (ResultSet resultSet, int rowNum) ->
-            Post.builder()
-                    .id(resultSet.getLong("id"))
-                    .memberId(resultSet.getLong("memberId"))
-                    .contents(resultSet.getString("contents"))
-                    .createdDate(resultSet.getObject("createdDate", LocalDate.class))
-                    .createdAt(resultSet.getObject("createdAt", LocalDateTime.class))
-                    .build();
-    private static final RowMapper<DailyPostCount> DAILY_POST_COUNT_MAPPER = (ResultSet resultSet, int rowNum) ->
-            new DailyPostCount(
-                    resultSet.getLong("memberId"),
-                    resultSet.getObject("createdDate", LocalDate.class),
-                    resultSet.getLong("count")
-            );
+public interface PostRepository {
+//    List<DailyPostCount> groupByCreatedDate(DailyPostCountRequest request);
+    List<Post> findAllByIdIn(List<Long> Ids);
+    Page<Post> findAllByMemberId(Long memberId, Pageable pageable);
+    Long countByMemberId(Long memberId);
+    List<Post> findAllByMemberIdOrderByIdDesc(Long memberId, Pageable pageable);
+    List<Post> findAllByMemberIdInOrderByIdDesc(List<Long> memberIds, Pageable pageable);
+    List<Post> findAllByIdLessThanAndMemberIdOrderByIdDesc(Long id, Long memberId, Pageable pageable);
+    List<Post> findAllByIdLessThanAndMemberIdInOrderByIdDesc(Long id, List<Long> memberIds, Pageable pageable);
+    Post save(Post post);
+//    List<Post> saveAll(List<Post> posts);
 
-    public List<DailyPostCount> groupByCreatedDate(DailyPostCountRequest request) {
-        String sql = String.format("""
-                SELECT createdDate, memberId, count(id) as count
-                FROM %s
-                WHERE memberId = :memberId
-                AND createdDate BETWEEN :firstDate and :lastDate
-                GROUP BY createdDate, memberId
-                """, TABLE);
-        SqlParameterSource params = new BeanPropertySqlParameterSource(request);
-        return namedParameterJdbcTemplate.query(sql, params, DAILY_POST_COUNT_MAPPER);
-    }
-
-    public List<Post> findAllbyInId(List<Long> Ids) {
-        if (Ids.isEmpty()) {
-            return List.of();
-        }
-        String sql = String.format("""
-                SELECT *
-                FROM %s
-                WHERE id in (:Ids)
-                """, TABLE);
-        SqlParameterSource params = new MapSqlParameterSource()
-                .addValue("Ids", Ids);
-        return namedParameterJdbcTemplate.query(sql, params, ROW_MAPPER);
-    }
-    public Page<Post> findAllByMemberId(Long memberId, Pageable pageable) {
-        String sql = String.format("""
-                SELECT *
-                FROM %s
-                WHERE memberId = :memberId
-                ORDER BY %s
-                LIMIT :size
-                OFFSET :offset
-                """, TABLE, PageHelper.orderBy(pageable.getSort()));
-        SqlParameterSource params = new MapSqlParameterSource()
-                .addValue("memberId", memberId)
-                .addValue("size", pageable.getPageSize())
-                .addValue("offset", pageable.getOffset());
-        List<Post> posts = namedParameterJdbcTemplate.query(sql, params, ROW_MAPPER);
-        return new PageImpl(posts, pageable, getCount(memberId));
-    }
-    private Long getCount(Long memberId) {
-        String sql = String.format("""
-                SELECT COUNT(*)
-                FROM %s
-                WHERE memberid = :memberId              
-                """, TABLE);
-        SqlParameterSource params = new MapSqlParameterSource().addValue("memberId", memberId);
-        return namedParameterJdbcTemplate.queryForObject(sql, params, Long.class);
-    }
-    public List<Post> findAllByMemberIdAndOrderByIdDesc(Long memberId, int size) {
-        String sql = String.format("""
-                SELECT *
-                FROM %s
-                WHERE memberId = :memberId
-                ORDER BY id desc
-                LIMIT :size
-                """, TABLE);
-        SqlParameterSource params = new MapSqlParameterSource()
-                .addValue("memberId", memberId)
-                .addValue("size", size);
-        return namedParameterJdbcTemplate.query(sql, params, ROW_MAPPER);
-    }
-    public List<Post> findAllByInMemberIdAndOrderByIdDesc(List<Long> memberIds, int size) {
-        if (memberIds.isEmpty()) {
-            return List.of();
-        }
-        String sql = String.format("""
-                SELECT *
-                FROM %s
-                WHERE memberId in (:memberIds)
-                ORDER BY id desc
-                LIMIT :size
-                """, TABLE);
-        SqlParameterSource params = new MapSqlParameterSource()
-                .addValue("memberIds", memberIds)
-                .addValue("size", size);
-        return namedParameterJdbcTemplate.query(sql, params, ROW_MAPPER);
-    }
-    public List<Post> findAllByLessThanIdAndMemberIdAndOrderByIdDesc(Long id, Long memberId, int size) {
-        String sql = String.format("""
-                SELECT *
-                FROM %s
-                WHERE memberId = :memberId and id < :id
-                ORDER BY id desc
-                LIMIT :size
-                """, TABLE);
-        SqlParameterSource params = new MapSqlParameterSource()
-                .addValue("memberId", memberId)
-                .addValue("id", id)
-                .addValue("size", size);
-        return namedParameterJdbcTemplate.query(sql, params, ROW_MAPPER);
-    }
-
-    public List<Post> findAllByLessThanIdAndInMemberIdAndOrderByIdDesc(Long id, List<Long> memberIds, int size) {
-        if (memberIds.isEmpty()) {
-            return List.of();
-        }
-        String sql = String.format("""
-                SELECT *
-                FROM %s
-                WHERE memberId in (:memberIds) and id < :id
-                ORDER BY id desc
-                LIMIT :size
-                """, TABLE);
-        SqlParameterSource params = new MapSqlParameterSource()
-                .addValue("memberIds", memberIds)
-                .addValue("id", id)
-                .addValue("size", size);
-        return namedParameterJdbcTemplate.query(sql, params, ROW_MAPPER);
-    }
-    public Post save(Post post) {
-        if (post.getId() == null) {
-            return insert(post);
-        }
-        throw new UnsupportedOperationException("Post는 갱신을 지원하지 않습니다.");
-    }
-    private Post insert(Post post) {
-        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(namedParameterJdbcTemplate.getJdbcTemplate())
-                .withTableName(TABLE)
-                .usingGeneratedKeyColumns("id");
-        SqlParameterSource params = new BeanPropertySqlParameterSource(post);
-        Long id = simpleJdbcInsert.executeAndReturnKey(params).longValue();
-        return Post
-                .builder()
-                .id(id)
-                .memberId(post.getMemberId())
-                .contents(post.getContents())
-                .createdDate(post.getCreatedDate())
-                .createdAt(post.getCreatedAt())
-                .build();
-    }
-    public void bulkInsert(List<Post> posts) {
-        String sql = String.format("""
-                INSERT INTO `%s` (memberId, contents, createdDate, createdAt)
-                VALUES (:memberId, :contents, :createdDate, :createdAt)
-                """, TABLE);
-        SqlParameterSource[] params = posts.stream()
-                .map(BeanPropertySqlParameterSource::new)
-                .toArray(SqlParameterSource[]::new);
-        namedParameterJdbcTemplate.batchUpdate(sql, params);
-    }
 }
