+++ src/main/java/com/example/fastcampusmysql/domain/post/service/PostReadService.java
@@ -3,6 +3,7 @@ package com.example.fastcampusmysql.domain.post.service;
 import com.example.fastcampusmysql.domain.post.dto.DailyPostCount;
 import com.example.fastcampusmysql.domain.post.dto.DailyPostCountRequest;
 import com.example.fastcampusmysql.domain.post.entity.Post;
+import com.example.fastcampusmysql.domain.post.repository.JpaPostRepository;
 import com.example.fastcampusmysql.domain.post.repository.PostRepository;
 import com.example.fastcampusmysql.util.CursorRequest;
 import com.example.fastcampusmysql.util.PageCursor;
@@ -10,16 +11,16 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
-import java.awt.*;
-import java.time.LocalDate;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class PostReadService {
-    private final PostRepository postRepository;
+//    private final PostRepository postRepository;
+    private final JpaPostRepository postRepository;
 
     public List<DailyPostCount> getDailyPostCount(DailyPostCountRequest request) {
         /*
@@ -29,7 +30,6 @@ public class PostReadService {
             group by createdDate, memberId
          */
         return postRepository.groupByCreatedDate(request);
-
     }
     public Page<Post> getPosts(Long memberId, Pageable pageable) {
         return postRepository.findAllByMemberId(memberId, pageable);
@@ -46,21 +46,27 @@ public class PostReadService {
         return new PageCursor<>(cursorRequest.next(nextKey), posts);
     }
     public List<Post> getPosts(List<Long> Ids) {
-        return postRepository.findAllbyInId(Ids);
+        return postRepository.findAllByIdIn(Ids);
     }
 
     private List<Post> findAllBy(Long memberId, CursorRequest cursorRequest) {
+        Pageable pageable = PageRequest.of(0, cursorRequest.size());
         if (cursorRequest.hasKey()) {
-            return postRepository.findAllByLessThanIdAndMemberIdAndOrderByIdDesc(cursorRequest.key(), memberId, cursorRequest.size());
+//            return postRepository.findAllByIdLessThanAndMemberIdOrderByIdDesc(cursorRequest.key(), memberId, cursorRequest.size());
+            return postRepository.findAllByIdLessThanAndMemberIdOrderByIdDesc(cursorRequest.key(), memberId, pageable);
         } else {
-            return postRepository.findAllByMemberIdAndOrderByIdDesc(memberId, cursorRequest.size());
+//            return postRepository.findAllByMemberIdOrderByIdDesc(memberId, cursorRequest.size());
+            return postRepository.findAllByMemberIdOrderByIdDesc(memberId, pageable);
         }
     }
     private List<Post> findAllBy(List<Long> memberIds, CursorRequest cursorRequest) {
+        Pageable pageable = PageRequest.of(0, cursorRequest.size());
         if (cursorRequest.hasKey()) {
-            return postRepository.findAllByLessThanIdAndInMemberIdAndOrderByIdDesc(cursorRequest.key(), memberIds, cursorRequest.size());
+//            return postRepository.findAllByIdLessThanAndMemberIdInOrderByIdDesc(cursorRequest.key(), memberIds, cursorRequest.size());
+            return postRepository.findAllByIdLessThanAndMemberIdInOrderByIdDesc(cursorRequest.key(), memberIds, pageable);
         } else {
-            return postRepository.findAllByInMemberIdAndOrderByIdDesc(memberIds, cursorRequest.size());
+//            return postRepository.findAllByMemberIdInOrderByIdDesc(memberIds, cursorRequest.size());
+            return postRepository.findAllByMemberIdInOrderByIdDesc(memberIds, pageable);
         }
     }
