+++ src/main/resources/application.properties
@@ -3,6 +3,9 @@ logging.level.org.springframework.jdbc.core = TRACE
 spring.datasource.username=oracle
 spring.datasource.url=jdbc:mysql://localhost:3308/fast_sns?rewriteBatchedStatements=true&characterEncoding=UTF-8&serverTimezone=Asia/Seoul
 spring.datasource.password=oracle
-# ??? ??? ????? ????(???? ???)
+
 spring.jpa.hibernate.naming.physical-strategy = org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-#server.port=8081
+
+# jpa query ??
+logging.level.org.hibernate.SQL=DEBUG
+logging.level.org.hibernate.type.descriptor.sql.BasicBinder=TRACE
